+++ packages/common/src/sites/HubSites.ts
@@ -333,6 +333,24 @@ export async function updateSite(
     getPropertyMap()
   );
   let modelToUpdate = mapper.entityToStore(site, currentModel);
+  // ============================================================
+  // Entity to Store can not handle scenarios where deep properties have been removed
+  // via a run-time migration.
+  // At this time we don't have a comprehensize solution for this, so we just remove
+  // the well-known-invalid properties from the model here.
+  // NOTES:
+  // * Add test logic into the `describe("updateSite removes properties:"...` section
+  // * We ignore coverage on the delete statements below because they
+  // use the optional property chaining, but we don't want to add a bunch of
+  // complex test cases just to satisfy the coverage tool.
+  // The tests do verify that the properties are removed.
+  // ============================================================
+
+  // Old telemetry props: migrated to correct property path in _migrateTelemetryConfig
+  /* istanbul ignore next */
+  delete modelToUpdate.data?.values?.telemetry;
+  /* istanbul ignore next */
+  delete modelToUpdate.item?.properties?.telemetry;
 
   // handle any domain changes
   await handleDomainChanges(modelToUpdate, currentModel, requestOptions);